Minor features, more documentation, bugfixes

+++ include/draw/drawdata.hpp
@@ -1,11 +1,30 @@
+/*
+ * pathBTSP is a tool to solve, approximate and draw instances of BTSPP,
+ * BTSP and TSP. Drawing is limited to euclidean graphs.
+ * Copyright (C) 2023 Jurek Rostalsky
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
 #pragma once
 
 #include <array>
 #include <vector>
 
-#include "draw/buffers.hpp"
+// graph library
+#include "graph.hpp"
 
-#include "graph/graph.hpp"
+#include "draw/buffers.hpp"
 
 #include "solve/definitions.hpp"
 
@@ -78,8 +97,8 @@ public:
   VertexOrder() : pInitialized(INITIAL_ACTIVENESS) {}
   ~VertexOrder() = default;
 
-  void updateOrder(const std::vector<unsigned int>& order, const ProblemType& type);
-  const std::vector<unsigned int>& operator[](const ProblemType type) const { return pVertexOrder[std::to_underlying(type)]; }
+  void updateOrder(const std::vector<size_t>& order, const ProblemType& type);
+  const std::vector<uint32_t>& operator[](const ProblemType type) const { return pVertexOrder[std::to_underlying(type)]; }
   bool initialized(const ProblemType& type) const { return pInitialized[std::to_underlying(type)]; }
 
 private:
