+++ src/draw/draw.cpp
@@ -1,3 +1,21 @@
+/*
+ * pathBTSP is a tool to solve, approximate and draw instances of BTSPP,
+ * BTSP and TSP. Drawing is limited to euclidean graphs.
+ * Copyright (C) 2023 Jurek Rostalsky
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
 #include "draw/draw.hpp"
 
 #include <array>
@@ -13,7 +31,8 @@
 #include "draw/shader.hpp"
 #include "draw/variables.hpp"
 
-#include "graph/graph.hpp"
+// graph library
+#include "graph.hpp"
 
 namespace drawing {
 static RGBA_COLOUR operator*(const RGBA_COLOUR& colour, const float fade) {
@@ -40,7 +59,7 @@ static void drawVertices(const ShaderProgram& drawCircles, const size_t numberOf
 
 static void drawPath(const ShaderProgram& drawPathSegments,
                      const std::shared_ptr<ShaderBuffer> shaderBuffer,
-                     const std::vector<unsigned int>& order,
+                     const std::vector<uint32_t>& order,
                      const float thickness,
                      const RGBA_COLOUR& colour) {
   shaderBuffer->bufferSubData(order);
@@ -78,11 +97,11 @@ static void drawGraph(const ShaderProgram& drawLine, const FloatVertices& floatV
 static void drawOpenEarDecomposition(const ShaderProgram& drawLine,
                                      const std::shared_ptr<DrawData> drawData,
                                      const graph::EarDecomposition& openEarDecomp) {
-  for (unsigned int i = 0; i < openEarDecomp.ears.size(); ++i) {
+  for (size_t i = 0; i < openEarDecomp.ears.size(); ++i) {
     const std::vector<size_t>& chain = openEarDecomp.ears[i];
     RGBA_COLOUR colour =
         drawData->appearance.colour[std::to_underlying(ProblemType::BTSP_approx)] * ((float) i / (openEarDecomp.ears.size() - 1));
-    for (unsigned int j = chain.size() - 1; j > 0; --j) {
+    for (size_t j = chain.size() - 1; j > 0; --j) {
       drawEdge(drawLine, drawData->floatVertices, graph::Edge{chain[j], chain[j - 1]}, 5.0f, colour);
     }
   }
